cloud_functions: move shared code to common
